finished main functionality
