--- flow_screen_components/navigationPath/force-app/main/default/lwc/jsNavigationPath/jsNavigationPath.js-meta.xml
@@ -1,13 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<LightningComponentBundle xmlns="http://soap.sforce.com/2006/04/metadata">
-    <apiVersion>50.0</apiVersion>
-    <isExposed>true</isExposed>
-    <targets>
-        <target>lightning__FlowScreen</target>
-    </targets>
-    <targetConfigs>
-        <targetConfig targets="lightning__FlowScreen" >
-            <property name="recordId" label="recordId" type="String" role="inputOnly" description="record"/>
-        </targetConfig>
-    </targetConfigs>
-</LightningComponentBundle>
