+++ flow_screen_components/navigationPath/force-app/main/default/lwc/navigationPath/navigationPath.js-meta.xml
@@ -7,7 +7,10 @@
     </targets>
     <targetConfigs>
         <targetConfig targets="lightning__FlowScreen" >
-            <property name="placeholder" label="placeholder" type="String" role="inputOnly" description="foo"/>
+            <property name="targetStringCollection" label="Target String Collection" type="String" role="inputOnly" description="These strings will be used both for what the user sees in the button bar and the value that is passed out of the button bar when the user clicks"/>
+            <property name="targetStringCSV" label="Target String CSV" type="String" role="inputOnly" description="These strings will be used both for what the user sees in the button bar and the value that is passed out of the button bar when the user clicks"/>
+            <property name="selectedTarget" label="Selected Target" type="String" role="outputOnly" description="The label of the button that the user clicked on"/>
+
         </targetConfig>
     </targetConfigs>
 </LightningComponentBundle>
